Speed up build (#32)
Without this PR, it takes about 65 seconds on my Orin Nano to run `make setup`, which was the preferred way to build the module. With this PR, you only need to run `make setup` once at the very beginning, and then running `make build` (or `make dist/module.tar.gz`, or even just `make`) goes about 25 seconds faster than it used to. It's still slower than I'd like, but the wait is more tolerable...

Tried both in the clone of the repo I've been developing on and a brand new clone, seems to work on both.


* add 'set -x' to build.sh, with a comment explaining all the options

* add a make target to just build, without setting up from scratch

* source the venv for all commands that use it

* the default target should be the .tar.gz file

* bugfix: construct the dist/ directory during setup even if we don't build everything then

* run 'make setup' in github CI actions

* rename build.sh -> setup.sh